requirements-test.txt
# Requirements for the test script
pysnmp==6.2.6
python-dotenv==1.0.0